chore(crates/rolldown): rename `finalizer` to `module_finalizers`
