+++ crates/rolldown/src/lib.rs
@@ -2,7 +2,7 @@ mod ast_scanner;
 mod bundler;
 mod bundler_builder;
 mod chunk_graph;
-mod finalizer;
+mod module_finalizers;
 mod module_loader;
 mod runtime;
 mod stages;
