+++ crates/rolldown/src/module_finalizers/finalizer_context.rs
