+++ cspell.json
@@ -46,6 +46,7 @@
     "bindingify",
     "binstall",
     "bools",
+    "Boshen",
     "Brooooooklyn",
     "canonicalize",
     "cdylib",
@@ -68,6 +69,7 @@
     "evanw",
     "execa",
     "Finalizer",
+    "finalizers",
     "globset",
     "hashbang",
     "hasher",
@@ -81,6 +83,7 @@
     "insta",
     "instanceof",
     "Itertools",
+    "iwanabethatguy",
     "izip",
     "jsonschema",
     "justfile",
@@ -132,21 +135,19 @@
     "tinybench",
     "transpiling",
     "treeshake",
+    "trivias",
+    "Trivias",
     "typeof",
     "underfin",
     "UNKEYED",
+    "usize",
+    "vecmap",
     "vite",
     "vitepress",
     "wasi",
     "WASI",
-    "Yinan",
-    "Yunfei",
-    "Trivias",
-    "vecmap",
-    "usize",
-    "trivias",
     "Xiangjun",
-    "iwanabethatguy",
-    "Boshen"
+    "Yinan",
+    "Yunfei"
   ]
 }
